Create README.md
This project have back-end and frond-end development and it is a good practice for CRUD operations
FRONT-END:
It created by react. I use Formik library for create for and yup library to validate form informations.For styling I use bootstrap library.
BACK-END 
Spring-boot framework is used for creation and Oracle SQL is the database. With @CrossOrigin Annotation on the controller class we connect it with front-end.